Initializes triple-canvases demo
We're going to face lots of difficulties here:
  - how do we keep track of changes between those contexts and sync
    what has to be synced without loosing efficiency?
  - how do we provide a great api for dealing with multiple contexts?

This commit starts adressing frontend issues, which are a less hard than
those other issues, which will involve a lot of api changes.

In summary:
  - changes css to support 3 canvases
  - starts to separate frontend code
  - creates a store file that will contain the most recent
    version of the app state.
