+++ example/03-curves/03-curves.js
@@ -12,42 +12,51 @@ import BasicMaterial from "utea/materials/BasicMaterial";
 import Line from "utea/geometries/Line";
 
 
-let pb = new PaintBoard(document.querySelector('canvas'));
+let pbOpen = new PaintBoard(document.querySelector('#canvas-open'), {
+  clearColor: [0.0, 0.1, 0.0, 1.0]
+});
+let pbClosed = new PaintBoard(document.querySelector('#canvas-closed'), {
+  clearColor: [0.0, 0.0, 0.1, 1.0]
+});
+let pb3d = new PaintBoard(document.querySelector('#canvas-3d'), {
+  clearColor: [0.1, 0.0, 0.0, 1.0]
+});
+
 let camera = new OrthographicCamera();
 let renderer = new Renderer(camera);
 let curve;
 
 camera.position = [0.0, 0.0, 1.0];
-pb.camera = camera;
+pbOpen.camera = camera;
 
-let grid = new Renderable(pb._gl, {
-  geometry: new PlaneGrid(pb._gl, 25.0),
-  material: new BasicMaterial(pb._gl, [0.3, 0.3, 0.3]),
+let grid = new Renderable(pbOpen._gl, {
+  geometry: new PlaneGrid(pbOpen._gl, 25.0),
+  material: new BasicMaterial(pbOpen._gl, [0.3, 0.3, 0.3]),
   drawMode: 'LINES'
 });
 grid.position = [0.0, 0.0, -1.0];
 grid.scale = [.1, .1, .1];
 
-let xAxis = new Renderable(pb._gl, {
-  geometry: new Line(pb._gl, [-1.0, 0.0, 0.0], [1.0, 0.0, 0.2]),
-  material: new BasicMaterial(pb._gl, [1.0, 0.0, 0.0]),
+let xAxis = new Renderable(pbOpen._gl, {
+  geometry: new Line(pbOpen._gl, [-1.0, 0.0, 0.0], [1.0, 0.0, 0.2]),
+  material: new BasicMaterial(pbOpen._gl, [1.0, 0.0, 0.0]),
   drawMode: 'LINES',
 });
 
-let yAxis = new Renderable(pb._gl, {
-  geometry: new Line(pb._gl, [0.0, -1.0, 0.0], [0.0, 1.0, 0.2]),
-  material: new BasicMaterial(pb._gl, [0.0, 1.0, 0.0]),
+let yAxis = new Renderable(pbOpen._gl, {
+  geometry: new Line(pbOpen._gl, [0.0, -1.0, 0.0], [0.0, 1.0, 0.2]),
+  material: new BasicMaterial(pbOpen._gl, [0.0, 1.0, 0.0]),
   drawMode: 'LINES',
 });
 
-let rags = new RaGs(pb._gl, camera, [
+let rags = new RaGs(pbOpen._gl, camera, [
   -0.5, 0.0, 0.0,
    0.0, 0.5, 0.0,
    0.5, 0.5, 0.0,
    0.5, 0.0, 0.0,
 ]);
 
-let nurbs = new NURBS(pb._gl, camera, new Float32Array([
+let nurbs = new NURBS(pbOpen._gl, camera, new Float32Array([
   -0.5, 0.0, 0.0,
    0.0, 0.5, 0.0,
    0.5, 0.5, 0.0,
@@ -55,152 +64,12 @@ let nurbs = new NURBS(pb._gl, camera, new Float32Array([
 ]));
 
 curve = nurbs;
-
-// TODO implement unproject inside camera base
-const unproject = (evt, pt) => {
-  vec3.copy(pt, [
-    2*evt.clientX/pb.width - 1,
-    1 - 2*evt.clientY/pb.height,
-    0.0
-  ]);
-
-  vec3.transformMat4(pt, pt, camera.inverseProjectionViewMatrix);
-  pt[2] = 0.0;
-};
-
-// TODO separate this UI stuff.
-//      We'll also need some kind of Store that
-//      will allow us to exchange data between
-//      different contexts and keep everything
-//      in sync.
-
-// GLOBALS
-let g_point = vec3.create();
-let g_editMode = 0x1; // 0 => insert | 1 => edit
-let g_selectedCp = -1;
-
-const ELEMS = {
-  bMode: document.querySelector("#b-mode"),
-  bClear: document.querySelector("b-clear"),
-  curveType: document.querySelector("#curve-type"),
-
-  widgetIterationsValue: document.querySelector(".widget.iterations span"),
-  widgetIterationsRange: document.querySelector(".widget.iterations input"),
-
-  widgetDegree: document.querySelector(".widget.degree"),
-  widgetDegreeRange: document.querySelector(".widget.degree input"),
-  widgetDegreeValue: document.querySelector(".widget.degree span"),
-
-  widgetVariance: document.querySelector(".widget.variance"),
-  widgetVarianceRange: document.querySelector(".widget.variance input"),
-  widgetVarianceValue: document.querySelector(".widget.variance span"),
-};
-
-ELEMS.bMode.addEventListener('click', (evt) => {
-  g_editMode ^= 0x1;
-
-  if (g_editMode)
-    evt.target.textContent = "INSERT";
-  else
-    evt.target.textContent = "EDIT";
-});
-
-ELEMS.curveType.addEventListener('change', (evt) => {
-  let tmpPoints = curve.points.control;
-  let tmpOffset = curve._offset;
-
-  switch (evt.target.value) {
-    case "rag":
-    ELEMS.widgetDegree.hidden = true;
-    ELEMS.widgetVariance.hidden = false;
-    curve = rags;
-    break;
-
-    case "nurbs":
-    ELEMS.widgetDegree.hidden = false;
-    ELEMS.widgetVariance.hidden = true;
-    curve = nurbs;
-    break;
-
-    default:
-      throw new Error("Unrecognized curve type selected.");
-  }
-
-  curve.setControlPoints(tmpPoints, tmpOffset);
-  draw();
-});
-
-ELEMS.widgetVarianceRange.addEventListener('input', (evt) => {
-  ELEMS.widgetVarianceValue.textContent = evt.target.value;
-  rags.variance = +evt.target.value;
-
-  draw();
-});
-
-ELEMS.widgetDegreeRange.addEventListener('input', (evt) => {
-  ELEMS.widgetDegreeValue.textContent = evt.target.value;
-  nurbs.degree = +evt.target.value;
-
-  draw();
-});
-
-ELEMS.widgetIterationsRange.addEventListener('input', (evt) => {
-  ELEMS.widgetIterationsValue.textContent = evt.target.value;
-  curve.iterations = evt.target.value;
-
-  draw();
-});
-
-pb.bindControls({
-  onMouseDown: (evt) => {
-    g_point = vec3.create();
-    unproject(evt, g_point);
-
-    if (!g_editMode) { // in insert mode
-      curve.addControlPoint(g_point);
-      ELEMS.widgetDegreeRange.max = curve._offset/3;
-      draw();
-
-      return;
-    }
-
-    g_selectedCp = curve.intersectsControlPoint(g_point);
-  },
-
-  onMouseMove: (evt) => {
-    if (!g_editMode)
-      return;
-
-    if (g_selectedCp < 0)
-      return;
-
-    unproject(evt, g_point);
-    curve.updateControlPoint(g_selectedCp, g_point);
-
-    if (~g_selectedCp)
-      draw();
-  },
-
-  onMouseUp: (evt) => {
-    if (!g_editMode)
-      return;
-
-    if (g_selectedCp < 0)
-      return;
-
-    unproject(evt, g_point);
-    curve.updateControlPoint(g_selectedCp, g_point);
-    g_selectedCp = -1;
-
-    draw();
-  },
-});
-
 renderer.submit(grid, xAxis, yAxis);
 
-
 function draw() {
-  pb.update();
+  pbClosed.update();
+  pb3d.update();
+  pbOpen.update();
   renderer.flush();
   curve.render();
 };
