+++ example/03-curves/index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width">
     <title>utea</title>
+    <link rel="stylesheet" href="../../vendor/reset.css">
     <link rel="stylesheet" href="../common.css" type="text/css">
     <link rel="stylesheet" href="03-curves.css" type="text/css">
   </head>
@@ -31,7 +32,9 @@
         </div>
       </div>
 
-      <canvas id="canvas"></canvas>
+      <canvas id="canvas-open"></canvas>
+      <canvas id="canvas-3d"></canvas>
+      <canvas id="canvas-closed"></canvas>
     <script src="../../03-curves.js"></script>
   </body>
 </html>
