+++ example/03-curves/store.js
@@ -1,0 +1,7 @@
+let curves = {
+  open: {current, rag, nurbs}
+  closed: {current, rag, nurbs}
+  surface: {}
+};
+
+
