+++ src/PaintBoard.js
@@ -2,7 +2,7 @@ if (__DEV__)
   var {createDebugContext} = require("./utils/debug.js");
 
 export default class PaintBoard {
-  constructor (canvas) {
+  constructor (canvas, opts={}) {
     this._gl;
     this._canvas = canvas;
     this._camera;
@@ -12,7 +12,11 @@ export default class PaintBoard {
 
     this._create3DContext();
     this._gl.enable(this._gl.DEPTH_TEST);
-    this._gl.clearColor(0.0, 0.0, 0.0, 1.0);
+
+    if (opts.clearColor)
+      this._gl.clearColor(...opts.clearColor);
+    else
+      this._gl.clearColor(0.0, 0.0, 0.0, 1.0);
   }
 
   get width () { return this._canvas.width; }
